backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
google-generativeai==0.8.3
pgvector==0.3.6
passlib[bcrypt]==1.7.4
bcrypt<5.0  # Compatibility fix for passlib
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
alembic==1.17.1

# Testing
pytest==8.3.5
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.15.1
faker==37.12.0
